[Event Hubs] Remove the middle layer of EventHubConsumer
